Extract native messaging types into their own files
