+++ apps/desktop/native-messaging-test-runner/src/bw-credential-update.ts
@@ -1,0 +1,60 @@
+// import "module-alias/register";
+
+// // import yargs from "yargs";
+// // import { hideBin } from "yargs/helpers";
+
+// import { LogUtils } from "./logUtils";
+// import NativeMessageService from "./nativeMessageService";
+// import * as config from "./variables";
+
+// // const argv: any = yargs(hideBin(process.argv)).option("name", {
+// //   alias: "n",
+// //   demand: true,
+// //   describe: "Name that the created login will be given",
+// //   type: "string",
+// // }).argv;
+
+// // const { name } = argv;
+
+// (async () => {
+//   const nativeMessageService = new NativeMessageService(1.0);
+//   // Handshake
+//   LogUtils.logWarning("Sending Handshake");
+//   const handshakeResponse = await nativeMessageService.sendHandshake(config.testRsaPublicKey);
+
+//   if (handshakeResponse.status !== "success") {
+//     LogUtils.logError(" Handshake failed. Status was:", handshakeResponse.status);
+//     nativeMessageService.disconnect();
+//     return;
+//   }
+//   LogUtils.logSuccess("Handshake success response");
+
+//   // Get active account userId
+//   const status = await nativeMessageService.checkStatus(handshakeResponse.sharedKey);
+
+//   const activeUser = status.payload.filter((a) => a.active === true && a.status === "unlocked")[0];
+//   if (activeUser === undefined) {
+//     LogUtils.logError("No active or unlocked user");
+//   }
+//   LogUtils.logWarning("Active userId: " + activeUser.id);
+
+//   const response = await nativeMessageService.credentialUpdate(
+//     handshakeResponse.sharedKey,
+//     "name",
+//     "",
+//     "",
+//     "",
+//     activeUser.id,
+//     ""
+//   );
+
+//   if (response.payload.status === "failure") {
+//     LogUtils.logError("Failure response returned ");
+//   } else if (response.payload.status === "success") {
+//     LogUtils.logSuccess("Success response returned ");
+//   } else {
+//     LogUtils.logWarning("Other response: ", response);
+//   }
+
+//   nativeMessageService.disconnect();
+// })();
