+++ apps/desktop/native-messaging-test-runner/src/ipcService.ts
@@ -4,7 +4,8 @@ import { homedir } from "os";
 
 import * as NodeIPC from "node-ipc";
 
-import { MessageCommon, UnencryptedMessageResponse } from "../../src/models/native-messages";
+import { UnencryptedMessageResponse } from "../../src/models/UnencryptedMessageResponse";
+import { MessageCommon } from "../../src/models/messageCommon";
 
 import Deferred from "./deferred";
 import { race } from "./race";
