+++ apps/desktop/src/models/decryptedCommandData.ts
@@ -1,0 +1,6 @@
+import { EncryptedCommand } from "./EncryptedCommand";
+
+export type DecryptedCommandData = {
+  command: EncryptedCommand;
+  payload?: any;
+};
