+++ apps/desktop/src/models/encryptedCommand.ts
@@ -1,0 +1,6 @@
+export type EncryptedCommand =
+  | "bw-status"
+  | "bw-credential-retrieval"
+  | "bw-credential-create"
+  | "bw-credential-update"
+  | "bw-generate-password";
