+++ apps/desktop/src/models/encryptedMessageResponse.ts
@@ -1,0 +1,7 @@
+import { EncString } from "@bitwarden/common/models/domain/encString";
+
+import { MessageCommon } from "./MessageCommon";
+
+export type EncryptedMessageResponse = MessageCommon & {
+  encryptedPayload: EncString;
+};
