+++ apps/desktop/src/models/legacyMessage.ts
@@ -1,0 +1,8 @@
+export type LegacyMessage = {
+  command: string;
+
+  userId?: string;
+  timestamp?: number;
+
+  publicKey?: string;
+};
