+++ apps/desktop/src/models/message.ts
@@ -1,0 +1,4 @@
+import { EncryptedMessage } from "./encryptedMessage";
+import { UnencryptedMessage } from "./unencryptedMessage";
+
+export type Message = UnencryptedMessage | EncryptedMessage;
