+++ apps/desktop/src/models/messageCommon.ts
@@ -1,0 +1,4 @@
+export interface MessageCommon {
+  version: number;
+  messageId: string;
+}
