+++ apps/desktop/src/models/unencryptedCommand.ts
@@ -1,0 +1,1 @@
+export type UnencryptedCommand = "bw-handshake";
