+++ apps/desktop/src/models/unencryptedMessage.ts
@@ -1,0 +1,9 @@
+import { MessageCommon } from "./MessageCommon";
+import { UnencryptedCommand } from "./UnencryptedCommand";
+
+export type UnencryptedMessage = MessageCommon & {
+  command: UnencryptedCommand;
+  payload: {
+    publicKey: string;
+  };
+};
