+++ apps/desktop/src/models/unencryptedMessageResponse.ts
@@ -1,0 +1,21 @@
+import { MessageCommon } from "./messageCommon";
+
+export type UnencryptedMessageResponse = MessageCommon &
+  (
+    | {
+        payload: {
+          status: "cancelled";
+        };
+      }
+    | {
+        payload: {
+          status: "success";
+          sharedKey: string;
+        };
+      }
+    | {
+        payload: {
+          error: "locked" | "cannot-decrypt";
+        };
+      }
+  );
