+++ apps/desktop/src/services/nativeMessageHandler.service.ts
@@ -3,21 +3,19 @@ import { ipcRenderer } from "electron";
 
 import { CryptoService } from "@bitwarden/common/abstractions/crypto.service";
 import { CryptoFunctionService } from "@bitwarden/common/abstractions/cryptoFunction.service";
-import { lockedUnlockedStatusString } from "@bitwarden/common/enums/authenticationStatus";
+import { AuthenticationStatus } from "@bitwarden/common/enums/authenticationStatus";
 import { Utils } from "@bitwarden/common/misc/utils";
 import { EncString } from "@bitwarden/common/models/domain/encString";
 import { SymmetricCryptoKey } from "@bitwarden/common/models/domain/symmetricCryptoKey";
 import { AuthService } from "@bitwarden/common/services/auth.service";
 import { StateService } from "@bitwarden/common/services/state.service";
 
-import {
-  Message,
-  UnencryptedMessage,
-  UnencryptedMessageResponse,
-  EncryptedMessage,
-  EncryptedMessageResponse,
-  DecryptedCommandData,
-} from "../models/native-messages";
+import { DecryptedCommandData } from "src/models/DecryptedCommandData";
+import { EncryptedMessage } from "src/models/EncryptedMessage";
+import { UnencryptedMessageResponse } from "src/models/UnencryptedMessageResponse";
+import { EncryptedMessageResponse } from "src/models/encryptedMessageResponse";
+import { Message } from "src/models/message";
+import { UnencryptedMessage } from "src/models/unencryptedMessage";
 
 const EncryptionAlgorithm = "sha1";
 
@@ -122,7 +120,7 @@ export class NativeMessageHandler {
             return {
               id: userId,
               email,
-              status: lockedUnlockedStatusString(authStatus),
+              status: authStatus === AuthenticationStatus.Unlocked ? "unlocked" : "locked",
               active: userId === activeUserId,
             };
           })
