+++ libs/common/src/enums/authenticationStatus.ts
@@ -3,12 +3,3 @@ export enum AuthenticationStatus {
   Locked = 1,
   Unlocked = 2,
 }
-
-export const lockedUnlockedStatusString = (status: AuthenticationStatus): string => {
-  switch (status) {
-    case AuthenticationStatus.Locked:
-      return "locked";
-    case AuthenticationStatus.Unlocked:
-      return "unlocked";
-  }
-};
